about halfway through code review of variables.py
